requirements.txt
javalang>=0.13.0
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.900
python-dotenv>=0.19.0
google-cloud-aiplatform>=1.36.0
vertexai>=0.0.1
google-generativeai
python-dotenv
httpx